requirements.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
pillow==10.3.0
pyotp==2.9.0
python-decouple==3.8
sqlparse==0.5.0
tzdata==2024.1
